start implementing separate java usb serial interface
